+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/differentialdrivebot/Robot.java
@@ -7,6 +7,7 @@ package edu.wpi.first.wpilibj.examples.differentialdrivebot;
 import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.XboxController;
+import edu.wpi.first.wpilibj.examples.differentialdrivebot.subsystems.Drivetrain;
 
 public class Robot extends TimedRobot {
   private final XboxController m_controller = new XboxController(0);
